Modify UserRepository.

+++ src/main/java/com/jd2/moviebase/model/Genre.java
@@ -20,7 +20,7 @@ public class Genre {
     private Long id;
 
     @Column(name = "tmdb_id", nullable = false)
-    private Integer tmdbId;
+    private Long tmdbId;
 
     @Column(nullable = false)
     private String name;
